Test rendering of stories
Swap to jest-fetch-mock from msw.

I was having a weird thing with msw, where when node-request-fetch intercepted my requests, it seemed to override my https request with http, and then an error was thrown for the msw server not supporting http.

This could have easily been user error on my part, but I couldn't track it down quickly, so I just switched to this much less elegant approach.